Split options premium pricer to two contract types
