+++ contracts/tests/TestOptionsPremiumPricerInNative.sol
@@ -1,9 +1,11 @@
 //SPDX-License-Identifier: GPL-3.0
 pragma solidity 0.7.3;
 
-import {OptionsPremiumPricer} from "../core/OptionsPremiumPricer.sol";
+import {
+    OptionsPremiumPricerInNative
+} from "../core/OptionsPremiumPricerInNative.sol";
 
-contract TestOptionsPremiumPricer is OptionsPremiumPricer {
+contract TestOptionsPremiumPricerInNative is OptionsPremiumPricerInNative {
     constructor(
         address _pool,
         address _volatilityOracle,
@@ -11,7 +13,7 @@ contract TestOptionsPremiumPricer is OptionsPremiumPricer {
         address _stablesOracle,
         address _nativeTokenOracle
     )
-        OptionsPremiumPricer(
+        OptionsPremiumPricerInNative(
             _pool,
             _volatilityOracle,
             _priceOracle,
@@ -68,30 +70,6 @@ contract TestOptionsPremiumPricer is OptionsPremiumPricer {
         }
     }
 
-    function testGetPremiumNativePairsInStables(
-        uint256 st,
-        uint256 expiryTimestamp,
-        bool isPut
-    ) external view returns (uint256 result, uint256 gas) {
-        bytes memory data =
-            abi.encodeWithSelector(
-                this.getPremiumNativePairsInStables.selector,
-                st,
-                expiryTimestamp,
-                isPut
-            );
-
-        uint256 startgas = gasleft();
-        (bool success, bytes memory returnData) =
-            address(this).staticcall(data);
-        gas = startgas - gasleft();
-
-        result = 0;
-        if (success) {
-            result = abi.decode(returnData, (uint256));
-        }
-    }
-
     function testGetOptionDelta(uint256 st, uint256 expiryTimestamp)
         external
         view
@@ -100,33 +78,6 @@ contract TestOptionsPremiumPricer is OptionsPremiumPricer {
         bytes memory data =
             abi.encodeWithSelector(
                 bytes4(keccak256(bytes("getOptionDelta(uint256,uint256"))),
-                st,
-                expiryTimestamp
-            );
-
-        uint256 startgas = gasleft();
-        (bool success, bytes memory returnData) =
-            address(this).staticcall(data);
-        gas = startgas - gasleft();
-
-        result = 0;
-        if (success) {
-            result = abi.decode(returnData, (uint256));
-        }
-    }
-
-    function testGetOptionDeltaNativePairs(uint256 st, uint256 expiryTimestamp)
-        external
-        view
-        returns (uint256 result, uint256 gas)
-    {
-        bytes memory data =
-            abi.encodeWithSelector(
-                bytes4(
-                    keccak256(
-                        bytes("getOptionDeltaNativePairs(uint256,uint256")
-                    )
-                ),
                 st,
                 expiryTimestamp
             );
